requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0

# Together AI SDK
together==1.2.0

# Cartesia AI SDK
cartesia==1.0.5

# Data Validation
pydantic>=2.6.3,<3.0.0
pydantic-settings>=2.1.0

# Token Counting
tiktoken==0.5.2

# Audio Processing
pydub==0.25.1
soundfile==0.12.1
numpy==1.24.3
ffmpeg-python==0.2.0
scipy==1.11.4

# Utilities
pathlib==1.0.1
typing-extensions==4.9.0
colorama==0.4.6
rich==13.7.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Optional: For development
pytest==7.4.3
black==23.12.1
mypy==1.8.0
